Disable rubygems-bundler when GEM_HOME overriden
Hem stores it's pre-installed gem bin files in a non-writable directory. The rubygems-bundler gem uses executable-hooks to try to overwrite the binaries, which fail. Projects when using overriden paths don't need rubygems-bundler because only the project's gems are stored in the path. Therefore the rubygems-bundler gem doesn't need to be installed in this situation.